consolidate print code between log and show
