+++ go/cmd/dolt/commands/merge.go
@@ -209,7 +209,7 @@ func (cmd MergeCmd) Exec(ctx context.Context, commandStr string, args []string, 
 					pager := outputpager.Start()
 					defer pager.Stop()
 
-					printCommitInfo(pager, 0, false, "auto", commit)
+					PrintCommitInfo(pager, 0, false, "auto", commit)
 				})
 			}
 		}
