+++ go/cmd/dolt/commands/revert.go
@@ -144,7 +144,7 @@ func (cmd RevertCmd) Exec(ctx context.Context, commandStr string, args []string,
 		pager := outputpager.Start()
 		defer pager.Stop()
 
-		printCommitInfo(pager, 0, false, "auto", commit)
+		PrintCommitInfo(pager, 0, false, "auto", commit)
 	})
 
 	return 0
