+++ integration-tests/bats/show.bats
@@ -49,7 +49,8 @@ assert_has_key_value() {
 
 @test "show: on initialized repo" {
     run dolt show
-    [ "$status" -eq "0" ]
+    echo $output
+    [ "$status" -eq 1 ]
     [[ "$output" =~ "Initialize data repository" ]] || false
 }
